Provide a Guardian to safeguard outgoing requests
Implements a number of SSRF protections by providing a Safe function
that can be hooked into a net.Dialer's Control method. Which protections
are enabled can be configured at the Guardian's creation time.

Based on conversation on the Go issue tracker the Go authors did not
envision the Control method being used for this purpose. However, there
is no other available method for doing so at a low enough level like the
Dialer to ensure that all outgoing connections, not just HTTP, would be
protected. Though it might not be their reason for having the Control
method, it works well enough for this purpose in practice and seem to be
how SSRF protection is implemented by many other Go projects too.