CI: Use a scoped PAT for the update flow
When GitHub Actions is the one making a commit or PR to the repo it
doesn't trigger any workflows. This is annoying because we want the
tests to run when ssrfgen proposes an update.

SSRFGEN_UPDATE_TOKEN is a scoped PAT with permissions on just this
repository to do the commits instead. That should cause CI to trigger.