CI: Add workflow to skip tests
This adds a workflow that's the inverse of the test.yaml workflow. Due
to path filtering, if a commit only touches README.md the PR will be
stuck forever as the workflow in test is required but will never
trigger. By defining a second workflow with the exact same name for the
job that ignores the paths the other workflow triggers on we can get the
status green and move on.