Rework how we do IPv6 blocking
Currently the prefix 2000::/3, so 2000:0000:0000:0000:0000:0000:0000:0000 -
3fff:ffff:ffff:ffff:ffff:ffff:ffff:ffff, is defined as the Global
Unicast range for IPv6. Anything outside that range is something we
inherently don't want to connect to. However, within that range a number
of special purpose subnets have also been carved out we also don't want
to connect to.

This reworks the logic to check membership of the global unicast range
first. If it doesn't match we can stop here. If it does fall within the
global unicast range we can then check for the few special purpose
ranges within it. This limits the size of the list we need to check for
IPv6 quite dramatically.